feat(JPA DAO): Basic CRUD implemented
Will be adding postman exception handling and response bodies

FIX: Deleted Redundant Classes

+++ server/my-family-ws/src/main/java/com/bytepushers/family/dao/UserDAO.java
@@ -1,6 +1,8 @@
 package com.bytepushers.family.dao;
 
 import com.bytepushers.family.model.User;
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
@@ -9,10 +11,10 @@ public interface UserDAO {
 
     // Standard CRUD operations
     User createUser(User user);
-    User findUserById(Integer id);
+    User findUserById(Long id);
     List<User> findAllUsers();
     void updateUser(User user);
-    void deleteUser(Integer id);
+    void deleteUser(Long id);
 
     User findUserByEmail(String email);
 }
